Created custom hook to get all books
