+++ src/components/books/EmptyShelf.module.css
@@ -29,12 +29,12 @@
     font-size: 1.2rem;
     padding-bottom: 1.5rem;
     font-weight: 400;
-    color: var(--dark-blue);
+    color: var(--med-grey);
 }
 
 /* Media Queries */
 
-@media (min-width:1024px) {
+@media (min-width:768px) {
     .empty-shelf {
         display: grid;
         grid-template-columns: 1fr 1fr;
